实现了基本功能，添加了OLED模块

+++ Encoder/Core/Src/trail.c
@@ -2,7 +2,7 @@
 
 // 获取循迹模块的状态
 float getTrailStatus() {
-    int thisTrailStatus = 0;
+    float thisTrailStatus = 0;
     if(L2==0 && L1==0 && center==1 && R1==0 && R2==0) thisTrailStatus = 0.0;  // 直行
     else if(L2==1 && L1==1 && center==0 && R1==0 && R2==0) thisTrailStatus = -4.5;  // 大幅度左转
     else if(L2==1 && L1==0 && center==0 && R1==0 && R2==0) thisTrailStatus = -4.0;  // 中幅度左转
